Improvements in Client<->Server comm

+++ org.smartsnip/src/org/smartsnip/shared/XComment.java
@@ -6,13 +6,13 @@ import com.google.gwt.user.client.rpc.IsSerializable;
 
 public class XComment implements IsSerializable {
 	public final String owner;
-	public final int snippet;
+	public final long snippet;
 	public final String message;
 	public final int positiveVotes;
 	public final int negativeVotes;
 	public final Date time;
 
-	public XComment(String owner, int snippet, String message, int positiveVotes, int negativeVotes, Date time) {
+	public XComment(String owner, long snippet, String message, int positiveVotes, int negativeVotes, Date time) {
 		this.owner = owner;
 		this.snippet = snippet;
 		this.message = message;
